ARROW-16590: Consolidate utilities for dealing with row-major data into arrow/compute/row
ARROW-16590: Moved GroupBy out of the kernels layer (api_aggregate) and into the exec layer (exec/aggregate).  Added some comments and renamed a few fields to adhere to style conventions.

ARROW-16590: Fix includes in benchmark to address GroupBy move
