+++ cpp/src/arrow/compute/CMakeLists.txt
@@ -71,3 +71,4 @@ add_arrow_benchmark(function_benchmark PREFIX "arrow-compute")
 add_subdirectory(kernels)
 
 add_subdirectory(exec)
+add_subdirectory(row)
