+++ cpp/src/arrow/compute/exec/key_hash.h
@@ -23,8 +23,8 @@
 
 #include <cstdint>
 
-#include "arrow/compute/exec/key_encode.h"
 #include "arrow/compute/exec/util.h"
+#include "arrow/compute/light_array.h"
 
 namespace arrow {
 namespace compute {
@@ -45,8 +45,8 @@ class ARROW_EXPORT Hashing32 {
   friend void TestBloomSmall(BloomFilterBuildStrategy, int64_t, int, bool, bool);
 
  public:
-  static void HashMultiColumn(const std::vector<KeyColumnArray>& cols,
-                              KeyEncoder::KeyEncoderContext* ctx, uint32_t* out_hash);
+  static void HashMultiColumn(const std::vector<KeyColumnArray>& cols, LightContext* ctx,
+                              uint32_t* out_hash);
 
   static Status HashBatch(const ExecBatch& key_batch, uint32_t* hashes,
                           int64_t hardware_flags, util::TempVectorStack* temp_stack,
@@ -157,8 +157,8 @@ class ARROW_EXPORT Hashing64 {
   friend void TestBloomSmall(BloomFilterBuildStrategy, int64_t, int, bool, bool);
 
  public:
-  static void HashMultiColumn(const std::vector<KeyColumnArray>& cols,
-                              KeyEncoder::KeyEncoderContext* ctx, uint64_t* hashes);
+  static void HashMultiColumn(const std::vector<KeyColumnArray>& cols, LightContext* ctx,
+                              uint64_t* hashes);
 
   static Status HashBatch(const ExecBatch& key_batch, uint64_t* hashes,
                           int64_t hardware_flags, util::TempVectorStack* temp_stack,
